split in inliner and optimization
